Implement message handlers and responses
Figured out how to receive events from the RTM api
currently, messages are filtered based on regular expression
and if the userID is mentioned, we send a response.
I learned that the RTM api only recieves events from channels that
the app has been added too. AND requies the channel:read scope